Working on common SuiteSettings component. Switched to using the component in ag_suite_settings.vue.
